Add sd_configcheck output to flare command
