+++ agent.py
@@ -39,12 +39,12 @@ from util import (
     get_hostname,
     Watchdog,
 )
-from utils.flare import configcheck, Flare
+from utils.flare import Flare
+from utils.configcheck import configcheck, sd_configcheck
 from utils.jmx import jmx_command
 from utils.pidfile import PidFile
 from utils.profile import AgentProfiler
-from utils.service_discovery.configcheck import sd_configcheck
-from utils.service_discovery.config_stores import get_config_store, TRACE_CONFIG
+from utils.service_discovery.config_stores import get_config_store
 from utils.service_discovery.sd_backend import get_sd_backend
 
 # Constants
@@ -401,19 +401,7 @@ def main():
 
     elif 'configcheck' == command or 'configtest' == command:
         configcheck()
-
-        if agentConfig.get('service_discovery', False):
-            # set the TRACE_CONFIG flag to True to make load_check_directory return
-            # the source of config objects.
-            # Then call load_check_directory here and pass the result to sd_configcheck
-            # to avoid circular imports
-            agentConfig[TRACE_CONFIG] = True
-            configs = {
-                # check_name: (config_source, config)
-            }
-            print("\nLoading check configurations...\n\n")
-            configs = load_check_directory(agentConfig, hostname)
-            sd_configcheck(agentConfig, configs)
+        sd_configcheck(agentConfig)
 
     elif 'jmx' == command:
         jmx_command(args[1:], agentConfig)
